test: split desktop contract tests
