+++ tests/FileLogDBContractTests.swift
@@ -1,0 +1,105 @@
+import Foundation
+
+// MARK: - file/log/db
+
+func runFileLogDBContractTests(testDir: String, fixturesDir: String, decoder: JSONDecoder) throws {
+    try testDbDescribe(fixturesDir: fixturesDir, decoder: decoder)
+    try testRemoteLogViewerPinCommandShape(testDir: testDir)
+}
+
+func testDbDescribe(fixturesDir: String, decoder: JSONDecoder) throws {
+    print("Test: db-describe.json")
+    print("----------------------")
+
+    let fixture = URL(fileURLWithPath: "\(fixturesDir)/db-describe.json")
+
+    guard FileManager.default.fileExists(atPath: fixture.path) else {
+        throw NSError(domain: "ContractTest", code: 20,
+            userInfo: [NSLocalizedDescriptionKey: "Fixture not found: db-describe.json"])
+    }
+
+    let data = try Data(contentsOf: fixture)
+
+    // Decode through CLIResponse wrapper (matches HomeboyCLI)
+    let result = try decoder.decode(CLIResponse<DbOutput>.self, from: data)
+
+    guard result.success else {
+        throw NSError(domain: "ContractTest", code: 21,
+            userInfo: [NSLocalizedDescriptionKey: "db-describe.json: success=false"])
+    }
+
+    guard let output = result.data else {
+        throw NSError(domain: "ContractTest", code: 22,
+            userInfo: [NSLocalizedDescriptionKey: "db-describe.json: data field is nil"])
+    }
+
+    print("[PASS] Parsed CLIResponse wrapper")
+    print("[PASS] Parsed DbOutput with projectId: \(output.projectId)")
+
+    // Validate stdout contains parseable table data
+    guard let stdout = output.stdout else {
+        throw NSError(domain: "ContractTest", code: 23,
+            userInfo: [NSLocalizedDescriptionKey: "db-describe.json: stdout is nil"])
+    }
+
+    guard let tableData = stdout.data(using: .utf8) else {
+        throw NSError(domain: "ContractTest", code: 24,
+            userInfo: [NSLocalizedDescriptionKey: "db-describe.json: stdout is not valid UTF-8"])
+    }
+
+    // Parse tables from stdout (matches DatabaseBrowserViewModel.parseTables)
+    let tables = try JSONDecoder().decode([WPTable].self, from: tableData)
+
+    print("[PASS] Parsed \(tables.count) tables from stdout")
+
+    // Validate table fields
+    for table in tables {
+        guard !table.Name.isEmpty else {
+            throw NSError(domain: "ContractTest", code: 25,
+                userInfo: [NSLocalizedDescriptionKey: "Table Name is empty"])
+        }
+    }
+    print("[PASS] All tables have Name field")
+
+    print("")
+    print("Database Summary:")
+    print("  Tables: \(tables.count)")
+    for table in tables {
+        print("    - \(table.Name) (\(table.Rows ?? "?") rows)")
+    }
+    print("")
+}
+
+func testRemoteLogViewerPinCommandShape(testDir: String) throws {
+    print("Test: Remote Log Viewer pin command shape")
+    print("-----------------------------------------")
+
+    let sourcePath = URL(fileURLWithPath: testDir)
+        .deletingLastPathComponent()
+        .appendingPathComponent("Homeboy/Extensions/RemoteLogViewer/RemoteLogViewerViewModel.swift")
+
+    let source = try String(contentsOf: sourcePath, encoding: .utf8)
+    let currentPinCommand = #"["project", "pin", "add", "--type", "log", "--tail", String(log.tailLines), projectId, log.path]"#
+    let currentTailUpdateCommand = #"["project", "pin", "add", "--type", "log", "--tail", String(lines), projectId, log.path]"#
+    let oldPositionalFirstCommand = #"["project", "pin", "add", projectId, log.path, "--type", "log", "--tail"#
+
+    guard source.contains(currentPinCommand) else {
+        throw NSError(domain: "ContractTest", code: 70,
+            userInfo: [NSLocalizedDescriptionKey: "RemoteLogViewer pin command does not use current homeboy project pin add syntax"])
+    }
+    print("[PASS] Pin command puts --type/--tail before project/path")
+
+    guard source.contains(currentTailUpdateCommand) else {
+        throw NSError(domain: "ContractTest", code: 71,
+            userInfo: [NSLocalizedDescriptionKey: "RemoteLogViewer tail update command does not preserve current pin syntax"])
+    }
+    print("[PASS] Tail-line update command preserves --tail before project/path")
+
+    guard !source.contains(oldPositionalFirstCommand) else {
+        throw NSError(domain: "ContractTest", code: 72,
+            userInfo: [NSLocalizedDescriptionKey: "RemoteLogViewer still contains old positional-first project pin add syntax"])
+    }
+    print("[PASS] Old positional-first log pin syntax is absent")
+
+    print("")
+}
